Create touchable chat boxes in chat list tab
Change the tab-one to chat list, and change the tab-two to setting.

Chat list tab is a page which shows all on going chat. Chats are
displayed as a list of boexes in the chat list. Other's name , the
latest message , the time of that message, and the porfile photo
are displayed on the chat box.

"ProfileAvatar" is the componemnt which displays user profile photo.

"Chat" is the chat bx component.

Seting tab is still a empty page, only showing text "Setting".

Use react native paper compnenets.