Refactor tab Icons to functions
Refactor tab layout icons to functions, so that the purpose of
the icon can be show on its function name.

Change Chats tab's icon and add user (plus) icon.

Make ChatsTabIcon and SettingTabIcon to be opacity when they are
not focused so that users can know which tab they are currently in.