Create input bar component in the bottom of chat window
Create input bar in the bottom of chat window.

In thie commit, input bar can only enter text. The feature of sending
messages haven't implement in this commit.

Input bar size and chat history position is responsive to the number
of text enter into the TextInput componnet. When the amount of text in
the TextInput is increase, the Input bar's height incrase, and the chat
messages slide upward so that the input bar won't overlap the messages
have been showed on the screen.

Voice message icon, sticker icon, and picture selector icon have been
inlcuded, but the fearures haven't been implemented yet in this commit.
These icon is unclickable for now.