Implement functionality that store sent message to local storage
By tapping enter/send, text message is stored into local storage.
At the same time, the state "messages" in chat window is updated,
hence the newly sent message will show up on the chat window.

Whenever the chat window is rendered, chat window will get messages
from local storage, and render all messages on the screen.

A chats storage and chats context are created to manage the rendering
of chatbox in chatListPage chats context will get all exist chats' ID
from storage, fetch the last message and timestamp of each chat, and
render a list of chatboxs.

When a new message is sent inside a chat, the chats context will be
updated, and the re-rendering of chatbox will be triggered. This can
make sure the latest message and timestamp is always show on the chatbox.

A mock authentication is created for testing purpose.