Adjuest the order of messages list in chat window
In chat-window.tsx, FlatList's property "inverted" is set to be true
so that the FlatList is rendered from bottom to the top, and the
bottom of the message list will be render on screen first when the chat
window is just opened.

Since the message list is rendered in a reverted order, the
fetchAllMessage() method is changed so that the order of messages
fetch from thie method is from the newest message to the oldest one.

InputBar styles has a subtle change so that the bottom of the message
list can align with the top of the input bar