Fix the InputBar style in Android
In android system, there is extra space under the InputBar, and
extra top padding in the textInput component.

To fix extra top padding in TextInput component, set the
padding-top to 0 if the platform is Android

To fix the extra space under the InputBar, set the behavior attribute
of KeyboardAvoidingView set to be "padding"for both IOS and Android,
and set the OffsetFooter height to 0 for Android.