Control the number of messages loaded at once
Fatcor out ChatList componnet from chat-window screen.

Add a messages context, which is a middle layer between the components
and the local storage. Messages context manages flow between components
and local storage.

ChatList component request a messages load to the mwssages context, and
the context will load the requested messages from local storage to the
context itself, and then the component can get the messages from context
to do rendering.

In this manner, the messages contexts can limit the number of message
loaded from the local storage, so that the number of data carrying
by the component's state can be reduced, and the number of requests
to local storage can be reduced as well.

Each time the chat window screen is existed, the loaded messages for
that screen will be reset to a default number to reduce the number
of data carrying by the context.

A newly send or received message is handed to the context, and the
context will update its loaded messages pool so that chat window can
render them out. The context will also store the new message to the
local storage.