Add friend-settings screen
Add a friend-setting screen, in which a clear chat history function
and a remove friend function is provided. Friend's avatar, user name,
and account Id is showed on this screen as well.

Clear chat history function and remove friend funtion have not been
implement yet, whcih will be done in the next commit.

App's main text change to chinese.