Complete clear chat history funstion
Complete clear chat history function. Chat history can be deteled
from lcoal storage by tapping the clear chat history option bar.

Set the root layout to hide the SplashScreen when the layout is mounted.
This is for trying to solve the issue of stucking at the SplashScreen
when launch the app by TestFlight