Fix message table creating error and a avatar error
Fix an error which caused the chat messages do not render on the
chat window of a friend who is the first time user send message to.

The error happend because the message object context haven't been
created at the monent the friend just added. The context odject only
create when the first message is send to the friend.

To fixing this error, create the message context object into the
context object list when the friend is added.

Another error is that the friend avatar change the default icon and
styles when the app is navigated to chat-window screen and friend
detail screen.

The error is caused by typos in some params which hold the avatar
data. The typo params make the value become undefined, hence params
was set to the default values.