Implement login and logout functionality
Add a login page, which asks for ID and passwrod. Account will be
login if id and password are matched.

Add a logout button on setting page, used for logout the account.

An authenticaiton storage is added for store the user info if a user
is logged in. The user info will be clear from storage if the user
logout the account.

An authentication api is added for password and ID matched. This
file takes in charge of authentication process. Right now it just
uses mock data for authenticaition checking.