Fix a variable naming issue and setup icon and splash page
The variable which used for carrying value of "account_id" in
authentication storage named as "id", which made the variable fail
to carry the account_id value, but become undefined instead. The
issue fixed by change the variable name to account_id in
authentication.storage.tsx

uploaded new app logo and splash page.