Change timestamp system to display local time
The timestamp generated by the local storage changed to unix
format timestamp. The time displayed on the ChatBox will use the
unix timestamp to generate the local data and time.

Date and time are also added under the message bubble.

In friends context, a new state name current_talking_friend_id is
added so that the application can know user is sending to and
receviving messages from which friend so that the only the ChatBox
of that friend will be re-rendered.