Make the app become compatible with server (Chat Engine) data
Reconstructe the storage type of authentication data and chats data
from SQLite to AsyncStorage.

Message data is still store in SQLite table in local storage, but
the properties of table changed.

Most of components' props are changed to fit with the data structure
of the server (Chat Engine)'s data

The app can consume mock server data to display info on the app
normally