Prevent duplicated load of loaded message
Add a last_read data in chat storage. last_read keeps track of the
last message id which has loaded in message storage.

The api function GetUnreadMessages() will only read mesasges with
 id smaller than last_read from server, so that the message get
from serevr is all unloaded message.

In addition, the Profile bar's subtitle is changed so that it will
display both username and ID if the contact_username property is
provided to the component. Otherwise, ID will be displayed only.