Implement send message functionality
When user send a message, the new message will be store in message
context with a -1 message id and a Unix timestamp as a pending
message. The pending message will show on the chat window with a
spinning loading icon.

Whenever the server send us back a new_message object via websocket,
the new_message object will be sent to the receiveMessage() method
in message context. This method will take the message object and
update the state of the pending message in context.

After the context is updated, the message object will be sent to
the local message storage.

At the same time, websocket context will check if this message is
sent from the user. If this new message is sent from the user,
the setLastRead method in chat context will be triggered, and the
last read message of the new message's chat will be updayted to
this message (the message sent by self is treated as read already
by default)