Fix error which cause no WebSocket connection just after login
Due to asynchronous nature of "user" state in authentication context
and is_messages_initialized in message context, websocket was not
able to get the up-to-date data from this two context, so that the
websocket connection cannot be established.

Use useRef to hold the data, and update the data by using useEffect
and set user and is_messages_initialized to be dependencies. In this
way, websocket context can get the newest state of these data.