Add image viewer and send multiple image functionalities
By tapping the image in the chat, an image viewer will pop up, and
the image can be zoom in and out in the image viwer.

In the process of sending image, right now user can select at most 9
images at a time, and all the selected image will be sent one by one
once the "add" button is tapped in the image picker.