Handle file loading and failed sent
Display a modal showing file is loading when the image picker is
loading the image or video from device

Show a red exclamation icon beside the message bubble when a message
is failed to send. Pree the icon and trigger resending of this
message.

Show a video play icon in the middle of the video thumbnail if the
message is a video.